refactor(serialization): extract shared ElementJsonConverter
Unify the duplicated FHIRPath-primitive -> JSON conversion that existed in both
SourceNodeInstanceFactory and the mapping engine's JsonNodeMutator.

- Add ElementJsonConverter (Ignixa.Serialization) with ToJsonValue(object) (the
  shared primitive contract) and ToJsonNode(IElement) (clone source-backed values,
  rebuild other complex values by element name).
- SourceNodeInstanceFactory delegates to it; gains float/DateTime/DateTimeOffset
  handling that only the mutator had.
- JsonNodeMutator.SerializePrimitive delegates to ElementJsonConverter.ToJsonValue
  (behavior-preserving). Its Type.IsCollection-aware complex-element walk is left
  in place by design - a different model from whole-object construction.

The mapping engine is NOT wired to IInstanceFactory: its object creation is
path-based mutation with its own StructureMap create semantics, distinct from
instance-selector construction. Only the leaf conversion is shared.

Verified: FML tests (539) and SourceNodeInstanceFactory tests (6) green.
